backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
pandas
requests
web3
eth-account
dashscope
